+++ Lib/ctypes/test/test_as_parameter.py
@@ -133,7 +133,7 @@ class BasicWrapTestCase(unittest.TestCase):
         f.argtypes = [c_longlong, MyCallback]
 
         def callback(value):
-            self.failUnless(isinstance(value, (int, long)))
+            self.failUnless(isinstance(value, (int, int)))
             return value & 0x7FFFFFFF
 
         cb = MyCallback(callback)
